+++ internal/resumer/boltdbresumer/spec.go
@@ -25,6 +25,7 @@ type Spec struct {
 	StopAfterDownload bool
 	StopAfterMetadata bool
 	CompleteCmdRun    bool
+	Sequential        bool
 	Version           int
 }
 
@@ -42,6 +43,7 @@ type jsonSpec struct {
 	StopAfterDownload bool
 	StopAfterMetadata bool
 	CompleteCmdRun    bool
+	Sequential        bool
 	Version           int
 
 	// JSON unsafe types
@@ -67,6 +69,7 @@ func (s Spec) MarshalJSON() ([]byte, error) {
 		StopAfterDownload: s.StopAfterDownload,
 		StopAfterMetadata: s.StopAfterMetadata,
 		CompleteCmdRun:    s.CompleteCmdRun,
+		Sequential:        s.Sequential,
 		Version:           s.Version,
 
 		InfoHash:  base64.StdEncoding.EncodeToString(s.InfoHash),
@@ -110,6 +113,7 @@ func (s *Spec) UnmarshalJSON(b []byte) error {
 	s.StopAfterDownload = j.StopAfterDownload
 	s.StopAfterMetadata = j.StopAfterMetadata
 	s.CompleteCmdRun = j.CompleteCmdRun
+	s.Sequential = j.Sequential
 	s.Version = j.Version
 	return nil
 }